requirements.txt
accelerate==1.1.0
av==15.1.0
bitsandbytes==0.49.2
datasets==2.20.0
decord==0.6.0
deepspeed==0.18.3
einops==0.8.1
flash_attn==2.5.8
ftfy==6.3.1
huggingface_hub==0.36.2
numpy==1.26.4
pandas==2.3.3
peft==0.17.1
Pillow==11.0.0
pysubs2==1.8.0
PyYAML==6.0.3
regex==2025.9.18
Requests==2.32.5
timm==1.0.20
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.54.0
